Cookers Completed
Mixers and Ovens now have methods which allow the player to insert and retrieve items from them. Additionally, they have a Step method that is called by a timer. This allows an item to be cooked over time.

Additionally, added a garbage disposal, which allows a player to discard unwanted items.

TODO:
* Create Cake object
* Create Fingers object
* Create IceCream object